+++ tests/flask/test_extensions.py
@@ -17,6 +17,7 @@ from flask_web3 import current_web3
 
 from consensys_utils.flask.extensions import initialize_extensions, \
     initialize_health_extension, initialize_swagger_extension, initialize_web3_extension
+from consensys_utils.flask.extensions.iterable import FlaskIterable
 from consensys_utils.flask.extensions.swagger import Swagger
 
 
@@ -62,6 +63,45 @@ def test_web3_extension(client, config):
         return jsonify(current_web3.eth.blockNumber)
 
     assert client.get('/test-web3').status_code == 200
+
+
+def test_iterable_extension(client, config):
+    mock_next = MagicMock()
+
+    class IteratorTest:
+        def __init__(self):
+            self.meter = 0
+
+        def set_config(self, config):
+            self.meter = config['meter']
+
+        def __iter__(self):
+            return self
+
+        def __next__(self):
+            if self.meter >= 6:
+                raise StopIteration
+            mock_next(self.meter)
+            self.meter += 1
+
+    config['meter'] = 2
+    FlaskIterable(IteratorTest, app=client.application)
+
+    assert hasattr(client.application, 'iterator')
+    assert hasattr(client.application, '__iter__')
+    assert hasattr(client.application, '__next__')
+
+    next(client.application)
+    mock_next.assert_called_once_with(2)
+
+    next(client.application)
+    mock_next.assert_any_call(3)
+
+    for _ in iter(client.application):
+        continue
+
+    mock_next.assert_any_call(4)
+    mock_next.assert_any_call(5)
 
 
 def test_custom_swagger_extension(client, config):